% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.r
\name{ABC_Parallel_JRNMM2_sigmuC}
\alias{ABC_Parallel_JRNMM2_sigmuC}
\title{ABC_Parallel_JRNMM2_sigmuC}
\usage{
ABC_Parallel_JRNMM2_sigmuC(refY, spec_refY, spx, T, h, M, w, startv, A, B,
  a, b, v0, r, vmax)
}
\value{
d, tilde.theta
}
\description{
Run 1 iteration from the sdbmsABC after sampling tilde.theta=(sigma,mu,C). Possibly faster code
}
