% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.r
\name{ABC_Parallel_JRNMM_sigmuC}
\alias{ABC_Parallel_JRNMM_sigmuC}
\title{ABC_Parallel_JRNMM_sigmuC}
\usage{
ABC_Parallel_JRNMM_sigmuC(in_refData, in_specRef, in_spx, in_T, in_h, in_M,
  in_w, A, B, a, b, v0, r, vmax)
}
\value{
d, tilde.theta
}
\description{
Run 1 iteration from the sdbmsABC after sampling tilde.theta=(sigma,mu,C)
}
