% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitting_JRNMM_Cpp.R
\name{Splitting_JRNMM_Cpp}
\alias{Splitting_JRNMM_Cpp}
\title{Splitting_JRNMM_Cpp}
\usage{
Splitting_JRNMM_Cpp(h, startv, grid, dm, cm, mu, C, A, B, a, b, v0, r,
  vmax)
}
\value{
the trajectory of X=(X1,...,X6)
}
\description{
Simulate a 6-dimensional trajectory from the JR-NMM
using Strang splitting
}
