% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitting_JRNMM_Cpp.R
\name{Splitting_JRNMM_output_Cpp}
\alias{Splitting_JRNMM_output_Cpp}
\title{Splitting_JRNMM_output_Cpp}
\usage{
Splitting_JRNMM_output_Cpp(h, startv, grid, dm, cm, mu, C, A, B, a, b, v0,
  r, vmax)
}
\value{
the trajectory Y=X2-X3
}
\description{
Simulate the output trajectory from the JR-NMM
using Strang splitting
}
